+++ fastplotlib/axes/__init__.py
@@ -1,0 +1,8 @@
+from ._axes import Grid, Grids, Ruler, Axes
+
+__all__ = [
+    "Grid",
+    "Grids",
+    "Ruler",
+    "Axes",
+]
